feat(fr): 拆分数据处理与布局逻辑
